+++ src/lib.rs
@@ -1,12 +1,12 @@
-mod config;
-mod errors;
-mod finder;
-mod pipeline;
-mod sanitize_policy;
-mod value;
+pub mod config;
+pub mod errors;
+pub mod pipeline;
+pub mod value;
+pub mod finder;
+pub mod sanitize_policy;
 
 pub use self::config::{CastType, Config};
-pub use self::errors::{ParseError, PipelineError, ValidationError};
 pub use self::finder::Finder;
 pub use self::pipeline::{Pipeline, Proc};
 pub use self::value::Value;
+pub use self::errors::*;